EEG Download Button & Irregular EEG Dashboard
Description: 
Feature request for Wavelit to allow for downloading of EEGs to work in tandem with the search capabilities of the ‘Patient Dashboard’ that has previously been released in Wavelit. Embed a dashboard containing EEGs that have been flagged as ‘abnormal’; provide a dashboard of those EEGs along w/metadata + ability to modify ML onsets.